Fixed Memory Bugs and Added Features
The program now works again, however it is laggy.

--- SharedConcurrent Queue Changes ---
Fixed the bug associated with SharedConcurrentQueue::peek, now it works
for all indices

Implemented SharedConcurrentQueue::wait_until_not_empty() which prevents
the cache from
retriving data until there is something to retrieve

--- Transformer Changes ---
Added a variable to keep track of the current processor, this will be
useful for chaining in the future

Cleaned up Transformer.cpp and simplified the processor chain, it only
works with one processor right now for testing.

--- Source Changes ---
We no longer check for mats to nullptr because it typically passes