Began Multithreading
I have began designing the multithreading architecture as defined below.

All processors run in individual threads to allow asynchronous image
processing, the Transformer class distributes the threads and begins
their execution when the processor is added to the transformer.

The architecture needs some more work as it is still a little buggy.

Thus, there is no longer a need for Transformer::processFrames.