Fixed Camera Output Bug -- Working Commit
I fixed the bug associated with the camera outputs occasionally
switching or showing the other's output. This was probably due to using
a peek instead of a pop to look at the input cache, thus I implemented a
function FrameCache::pop to pop and return the top of the cache.